chore: move backup to StatementExecutor
